Issue #3 Fixed and Added some supporting functions for aiming
Fixed the recursive call issue, was using an ambiguous name for a function which was not caught by compiler. Removed the ambiguous function call, and replaced with the PlayerController's   GetViewportSize(int32&, int32&) function. 

Added Line tracing for player to get location of where  to aim to based on the screen location of the aim cross hair (dot for this version).

Added naive aim at for AIController, that takes the absolute player location and aims at it